Moved code for getting a topic's denormalized root content item from SlidePreviewsSidebar into Slides component.
