+++ app/core-components/slides/tests/Slides.test.js
@@ -4,36 +4,19 @@ import * as React from 'react';
 import { shallow } from 'enzyme';
 
 import { dummyTranslatorProps } from 'config/tests';
-import contentItems from 'modules/contentItems';
+import { dummyContentItemData as dummyData } from 'lib/testResources';
 
 import { PureSlides } from '../Slides';
-
-const { contentItemTypes } = contentItems.model;
 
 describe(`Slides`, (): void => {
 
   it(`renders without errors`, (): void => {
-    const dummyContentItemTreeRoot = {
-      id: 'abcdefghij',
-      type: contentItemTypes.ROOT,
-      isEditing: false,
-      childItemIds: ['hd83riji8c'],
-      childItems: [{
-        id: 'hd83riji8c',
-        type: contentItemTypes.HEADING,
-        isEditing: false,
-        text: 'Heading',
-        metadata: {
-          tags: [],
-          visibilityOverrides: {},
-        },
-        subItems: [],
-      }],
-    };
+    const dummyContentItemTreeRoot = { ...dummyData.rootContentItem, childItems: [] };
 
     const enzymeWrapper = shallow(
       <PureSlides
         {...dummyTranslatorProps}
+        topicId="dummyTopicId"
         contentItemTreeRootItem={dummyContentItemTreeRoot}
       />,
     );
