+++ app/modules/platform/components/Sidebars/sidebarIdsToComponentsMap/SlidePreviewsSidebar.test.js
@@ -4,23 +4,17 @@ import * as React from 'react';
 import { shallow } from 'enzyme';
 
 import { dummyTranslatorProps } from 'config/tests';
-import { dummyTopicData, dummyContentItemData } from 'lib/testResources';
+import { dummyTopicData } from 'lib/testResources';
 import topics from 'modules/topics';
-import contentItems from 'modules/contentItems';
 
 import { PureSlidePreviewsSidebar } from './SlidePreviewsSidebar';
 
 describe(`SlidePreviewsSidebar`, (): void => {
 
   let dummyTopic: topics.model.Topic;
-  let dummyContentItemTreeRootItem: contentItems.model.DenormalizedRootContentItem;
 
   beforeEach((): void => {
     dummyTopic = { ...dummyTopicData.topic };
-    dummyContentItemTreeRootItem = {
-      ...dummyContentItemData.rootContentItem,
-      childItems: [],
-    };
   });
 
   it(`renders without errors`, (): void => {
@@ -28,7 +22,6 @@ describe(`SlidePreviewsSidebar`, (): void => {
       <PureSlidePreviewsSidebar
         {...dummyTranslatorProps}
         topic={dummyTopic}
-        contentItemTreeRootItem={dummyContentItemTreeRootItem}
       />,
     );
     expect(enzymeWrapper.isEmptyRender()).toEqual(false);
